Reorganize and direct user to result page after job is created
