+++ src/App.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
-import './App.css';
-import Home from './pages/home';
-import Result from './pages/result';
-import Create from './pages/create';
+
+import Home from './components/home';
+import Result from './components/result';
+import CreateJob from './components/create-job';
 
 class App extends Component {
   render() {
@@ -12,7 +12,7 @@ class App extends Component {
         <Route path="/" exact component={Home} />
         <Route path="/result/" exact component={Result} />
         <Route path="/result/:resultId" component={Result} />
-        <Route path="/create/" component={Create} />
+        <Route path="/create/" component={CreateJob} />
       </Router>
     );
   }
