+++ src/components/result/index.jsx
