--- src/pages/create/index.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Create = props => {
-  return <div>Welcome to Create page</div>;
-};
-
-export default Create;
