--- src/pages/home/index.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Home = props => {
-  return <div>Welcome Home</div>;
-};
-
-export default Home;
